refactor(git-history): View 抽 branchDiff 段、Impl scope 类型迁出破倒置并抽数据加载段（View 3425→2988，Impl 4933→3662）
